+++ viewmodel/src/main/java/com/baghdad/viewmodel/movieDetails/MovieDetailsMapper.kt
@@ -1,8 +1,2 @@
 package com.baghdad.viewmodel.movieDetails
 
-import com.baghdad.domain.model.MediaAccountStates
-
-fun MediaAccountStates.toUiState() =
-    MovieDetailsState(
-        isRated = isMediaRated,
-    )
