+++ viewmodel/src/main/java/com/baghdad/viewmodel/review/ReviewMapper.kt
@@ -1,7 +1,7 @@
 package com.baghdad.viewmodel.review
 
 import com.baghdad.entity.media.Review
-import com.baghdad.viewmodel.movieDetails.roundToFirstDecimal
+import com.baghdad.viewmodel.util.roundToFirstDecimal
 import com.baghdad.viewmodel.util.toDDMMYYYYFormat
 
 fun Review.toReviewUi() = ReviewScreenState.ReviewUiState(
